Refactor: Move fileObjectWalker to dedicated file
